Add security certificates section below case studies section in the homepage
