+++ src/components/Homepage/SectionFive/index.tsx
@@ -1,23 +1,58 @@
+import { StaticImage } from 'gatsby-plugin-image';
 import * as React from 'react';
-import DarkSwoopingLinesLeftDirectionBackground from '../../BackgroundImages/DarkSwoopingLinesLeftDirectionBackground';
-import FlowDemoVideo from '../../FlowDemoVideo';
-import { Description, Title, Wrapper } from './styles';
+import RealTimeCdc from '../../../svgs/real-time-cdc.svg';
+import Advantage from './Advantage';
+import {
+    AdvantagesList,
+    Container,
+    Description,
+    IconWrapper,
+    LeftColumn,
+    Link,
+    RightColumn,
+    Title,
+    Wrapper,
+} from './styles';
 
 const SectionFive = () => {
     return (
-        <DarkSwoopingLinesLeftDirectionBackground>
+        <Container>
             <Wrapper>
-                <Title>
-                    <span>Build </span>
-                    in minutes
-                </Title>
-                <Description>
-                    Use the UI to quickly build end-to-end pipelines using
-                    no-code connectors. Estuary Flow does the rest.
-                </Description>
-                <FlowDemoVideo />
+                <LeftColumn>
+                    <IconWrapper>
+                        <RealTimeCdc />
+                    </IconWrapper>
+                    <Title>
+                        <span>Use the best </span>
+                        real-time CDC
+                    </Title>
+                    <Description>
+                        Estuary Flow is the most real-time, most reliable change
+                        data capture (CDC) available today. It is the only CDC
+                        with the:
+                    </Description>
+                    <AdvantagesList>
+                        <Advantage title="Fastest captures, with sub-100ms end-to-end latency" />
+                        <Advantage title="Most reliable delivery via stream-store-replay" />
+                        <Advantage title="Most flexible materializations that run at your speed of choice" />
+                        <Advantage title="Most automated pipelines with schema evolution" />
+                        <Advantage title="Most versatile writes that maintain a current view or all change history" />
+                        <Advantage title="Most scalable pipelines, with true elastic scaling" />
+                    </AdvantagesList>
+                    <Link href="/integrations" target="_blank" rel="noopener">
+                        View Connectors
+                    </Link>
+                </LeftColumn>
+                <RightColumn>
+                    <StaticImage
+                        src="../../../images/real-time-graphic.png"
+                        alt="Real-time ETL with Estuary Flow: Seamlessly move data from source to destination for immediate analysis and actionable insights."
+                        placeholder="none"
+                        quality={100}
+                    />
+                </RightColumn>
             </Wrapper>
-        </DarkSwoopingLinesLeftDirectionBackground>
+        </Container>
     );
 };
