+++ src/components/Homepage/SectionSix/styles.ts
@@ -1,44 +1,27 @@
 import styled from 'styled-components';
 import { globalMaxWidth, sectionTopBottomPadding } from '../../../globalStyles';
-import { OutboundLinkFilled } from '../../OutboundLink';
-
-export const Container = styled.section`
-    background-color: #fdfdfe;
-`;
 
 export const Wrapper = styled.div`
     ${globalMaxWidth}
     ${sectionTopBottomPadding}
-    
-  display: flex;
-    justify-content: space-between;
+
+    display: flex;
+    justify-content: center;
     align-items: center;
-    gap: 60px;
-
-    @media (max-width: 1024px) {
-        flex-direction: column;
-        align-items: flex-start;
-    }
-
-    @media (max-width: 425px) {
-        gap: 40px;
-    }
+    flex-direction: column;
 `;
 
 export const Title = styled.h2`
+    text-align: center;
     font-size: 2.5rem;
     font-weight: 600;
     line-height: 48px;
     color: #5072eb;
-    margin: 0;
     text-transform: uppercase;
+    margin: 0;
 
-    & span:nth-child(odd) {
-        color: #04192b;
-    }
-
-    @media (max-width: 1024px) {
-        font-size: 2.25rem;
+    span {
+        color: #ffffff;
     }
 `;
 
@@ -46,8 +29,10 @@ export const Description = styled.p`
     font-size: 1.25rem;
     font-weight: 400;
     line-height: 30px;
-    margin: 0;
-    color: #47506d;
+    color: #ffffff;
+    margin: 24px 0 60px 0;
+    max-width: 720px;
+    text-align: center;
 
     @media (max-width: 1024px) {
         font-size: 1rem;
@@ -56,37 +41,4 @@ export const Description = styled.p`
     @media (max-width: 980px) {
         max-width: 480px;
     }
-`;
-
-export const Button = styled(OutboundLinkFilled)`
-    padding: 16px 24px;
-    width: fit-content;
-
-    @media (max-width: 425px) {
-        width: 100%;
-    }
-`;
-
-export const IconWrapper = styled.div`
-    padding: 18px;
-    border: 1px solid #d7dce5;
-    border-radius: 8px;
-    width: fit-content;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #ffffff;
-`;
-
-export const LeftColumn = styled.div`
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-`;
-
-export const RightColumn = styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
 `;
