--- src/components/Homepage/index.tsx
@@ -1,23 +1,0 @@
-import SectionEight from './SectionEight';
-import SectionFive from './SectionFive';
-import SectionFour from './SectionFour';
-import SectionNine from './SectionNine';
-import SectionOne from './SectionOne';
-import SectionSeven from './SectionSeven';
-import SectionSix from './SectionSix';
-import SectionTen from './SectionTen';
-import SectionThree from './SectionThree';
-import SectionTwo from './SectionTwo';
-
-export {
-    SectionEight,
-    SectionFive,
-    SectionFour,
-    SectionNine,
-    SectionOne,
-    SectionSeven,
-    SectionSix,
-    SectionTen,
-    SectionThree,
-    SectionTwo,
-};
