+++ src/components/Product/SectionTen/index.tsx
@@ -1,36 +1,11 @@
-import { StaticImage } from 'gatsby-plugin-image';
 import React from 'react';
 import { DefaultWrapperDark } from '../../../styles/wrappers';
-import Advantages from '../../Advantages';
-import { sectionTenAdvantages } from '../advantages';
+import SecurityCertificates from '../../SecuriityCertificates';
 
 const SectionTen = () => {
     return (
         <DefaultWrapperDark>
-            <Advantages
-                icon={
-                    <StaticImage
-                        placeholder="none"
-                        alt="Secure Your Data with Estuary Flow: GDPR, CCPA, CPRA Compliant & SOC 2 Type II Certified"
-                        src="../../../svgs/product-page/section-ten/secure-data.svg"
-                    />
-                }
-                title={
-                    <>
-                        <span>SECURE</span> YOUR DATA
-                    </>
-                }
-                subtitle="Estuary Flow is designed and tested to make sure your data and your systems stay secure."
-                image={
-                    <StaticImage
-                        placeholder="none"
-                        alt="Secure Your Data with Estuary Flow: Compliance and Certification"
-                        src="../../../images/product-page/section-ten/gdpr-ccpa-cpra-soc2-type2-compliant.png"
-                        quality={100}
-                    />
-                }
-                advantages={sectionTenAdvantages}
-            />
+            <SecurityCertificates />
         </DefaultWrapperDark>
     );
 };
